Rename NonStaticBaseModule and delete BaseModule
