+++ packages/react-native-executorch/src/hooks/computer_vision/useClassification.ts
@@ -1,5 +1,5 @@
 import { ResourceSource } from '../../types/common';
-import { useNonStaticModule } from '../useNonStaticModule';
+import { useModule } from '../useModule';
 import { ClassificationModule } from '../../modules/computer_vision/ClassificationModule';
 
 interface Props {
@@ -8,7 +8,7 @@ interface Props {
 }
 
 export const useClassification = ({ model, preventLoad = false }: Props) =>
-  useNonStaticModule({
+  useModule({
     module: ClassificationModule,
     model,
     preventLoad: preventLoad,
