+++ packages/react-native-executorch/src/hooks/computer_vision/useImageSegmentation.ts
@@ -1,4 +1,4 @@
-import { useNonStaticModule } from '../useNonStaticModule';
+import { useModule } from '../useModule';
 import { ImageSegmentationModule } from '../../modules/computer_vision/ImageSegmentationModule';
 import { ResourceSource } from '../../types/common';
 
@@ -8,7 +8,7 @@ interface Props {
 }
 
 export const useImageSegmentation = ({ model, preventLoad = false }: Props) =>
-  useNonStaticModule({
+  useModule({
     module: ImageSegmentationModule,
     model,
     preventLoad,
