+++ packages/react-native-executorch/src/modules/natural_language_processing/TextEmbeddingsModule.ts
@@ -1,8 +1,8 @@
 import { ResourceSource } from '../../types/common';
 import { ResourceFetcher } from '../../utils/ResourceFetcher';
-import { BaseNonStaticModule } from '../BaseNonStaticModule';
+import { BaseModule } from '../BaseModule';
 
-export class TextEmbeddingsModule extends BaseNonStaticModule {
+export class TextEmbeddingsModule extends BaseModule {
   async load(
     model: { modelSource: ResourceSource; tokenizerSource: ResourceSource },
     onDownloadProgressCallback: (progress: number) => void = () => {}
